+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Added
 - **Irodsbackend**
     - Support for ``IRODS_ROOT_PATH`` setting (#1067)
     - ``get_root_path()`` and ``get_projects_path()`` helpers (#1067)
+    - Optional ``user_name`` and ``user_pass`` in ``IrodsAPI`` init kwargs (#1139)
 - **Samplesheets**
     - Warning for unrecognized assay plugin in sample sheet import (#1070)
     - Sheet creation from templates using cubi-tk (#1068)
@@ -30,6 +31,7 @@ Added
     - Enable remote sync for sample sheets (#959, #1102, #1103)
     - ``Icon`` component in vue app for django-iconify icon access (#1113)
     - App alerts for iRODS data request actions (#1084)
+    - Public access support for sample data (#1100)
 
 Changed
 -------